+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## In Progress
 
+- `device` is now `endpoint`.
 - New comment syntax: '#' or '//' which consumes to EOL. The old 'note:'
   and 'comment:' syntax is removed.